+++ inc/timestamp.h
@@ -11,7 +11,7 @@ class timestamp{
 
 	short			ts_month, ts_hour, ts_minute;
 
-	float			ts_day, ts_second;
+	double			ts_day, ts_second;
 
 	bool			ts_isLeap;
 	bool			ts_validDate, ts_gregorianDate;
@@ -32,14 +32,14 @@ protected:
 public:
 
 	timestamp(void);
-	timestamp(int _year, short _month, float _day, short _hour, short _minute, float _second, bool _verbose);
+	timestamp(int _year, short _month, double _day, short _hour, short _minute, double _second, bool _verbose);
 
 	int		ts_getYear(void)			{return ts_year;}
 	short	ts_getMonth(void)			{return ts_month;}
-	float	ts_getDay(void)				{return ts_day;}
+	double	ts_getDay(void)				{return ts_day;}
 	short	ts_getHour(void)			{return ts_hour;}
 	short	ts_getMinute(void)			{return ts_minute;}
-	float	ts_getSecond(void)			{return ts_second;}
+	double	ts_getSecond(void)			{return ts_second;}
 	bool	ts_getLeap(void)			{return ts_isLeap;}
 	bool	ts_getGregorianDate(void)	{return ts_gregorianDate;}
